pyproject.toml
[project]
name = "shopping-agent"
version = "0.1.0"
description = "Shopping agent with Chainlit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "chainlit>=1.0.0",
  "openai-agents>=0.2.1",
  "python-dotenv>=1.1.1",
  "requests>=2.31.0"
]

[tool.chainlit]
project_root = "."
